Incorporate ~/.search-kit/config.yml settings.
This commit introduces a little more elaborate configuration process for
SearchKit.  Namely we ought to be able to setup SearchKit with an access
token and a default URI.

Things left to do for pleasant first run experience:

* Prompt for setup on first run (ask for configuration token)
* Refuse access to APIs kit-wide without token configuration

This begs for an authentication feature to be added to the search
process.  The infrastructure for authentication exists but the APIs
themselves are not currently using it.  This needs to change on the API
itself.