Support all current APIs.
This commit adds support for all initial APIs which search kit offers.
The CLI isn't yet fully-represented, but the internal interactors that
power it are.

Ideally, the CLI should both supply an agnostic, vanilla interface to
the client software, and a more human usable version.  Though both would
be useful for use and reuse, a human usable layer of the interface could
go far in making people enjoy using (and relying on) the CLI.