Add subscribers CLI and first run experience.
The subscribers API has been interfaced with on the command line.
Additionally, a new root command, `setup`, has been added to the CLI.
This prompts for a username and password, creates a subscriber account
and sets up the app token value for the CLI.

In this commit, it became obvious that the CLI autoloader was producing
some strange side-effects in terms of subcommand namespaces.  This needs
to be corrected.