More resilient logging behavior. Semvar bump 0.0.4
There was a bug going around in which Search Kit was attempting to log
in directories that didn't yet exist.  This commit ensures that the
directory exists, and also that it is established in a canonical
location (the gem's log folder, unless the logger is reassigned at
runtime).

This also marks release version 0.0.4